Patch for accepting more then 255 datastore connections

+++ src/dyn_conf.h
@@ -128,7 +128,7 @@ struct conf_pool {
   struct conf_listen stats_listen; /* stats_listen: socket info for stats */
 
   /* connection pool details */
-  uint8_t datastore_connections;
+  uint16_t datastore_connections;
   uint8_t local_peer_connections;
   uint8_t remote_peer_connections;
 };
